add pec rhel nic
