+++ spec/lib/coordinate/nic/ubuntu_spec.rb
@@ -1,9 +1,10 @@
 require 'spec_helper'
+require 'support/nic_helper'
 require 'ostruct'
 
 describe Pec::Coordinate::UserData::Nic::Ubuntu do
 
-  subject { Pec::Coordinate::UserData::Nic::Ubuntu.gen_user_data(networks, ports) }
+  subject { described_class.gen_user_data(networks, ports) }
   it {
     expect(subject).to eq(
       {
@@ -14,52 +15,4 @@ describe Pec::Coordinate::UserData::Nic::Ubuntu do
       }
     )
   }
-end
-
-def ports
-  [
-    OpenStruct.new({
-      id: 1,
-      name: "eth0",
-      mac_address: '00:00:00:00:00:00',
-      fixed_ips: [
-        {
-          'ip_address' => "10.10.10.10"
-        }
-      ]
-    }),
-    OpenStruct.new({
-      id: 1,
-      name: "eth1",
-      mac_address: '00:00:00:00:00:00',
-      fixed_ips: [
-        {
-          'ip_address' => "20.20.20.20"
-        }
-      ]
-    })
-  ]
-end
-
-def networks
-  [
-    [
-      "eth0",
-      {
-        "bootproto" => "static",
-        "ip_address" => "10.10.10.10/24",
-        "dns-nameservers" => "8.8.8.8",
-        "gateway" => "1.1.1.1"
-      }
-    ],
-    [
-      "eth1",
-      {
-        "bootproto" => "static",
-        "ip_address" => "20.20.20.20/24",
-        "dns-nameservers" => "8.8.8.8",
-        "gateway" => "2.2.2.2"
-      }
-    ]
-  ]
 end
